Modularize components, remove uneeded icons, edit local gitignore, remove local readme
