+++ src/frontend/.gitignore
@@ -12,6 +12,9 @@ dist
 dist-ssr
 *.local
 
+**/.gitignore
+!/.gitignore
+
 # Editor directories and files
 .vscode/*
 !.vscode/extensions.json
